refactor: class component ResultsBlock rename functional component
