+++ src/components/ResultsBlock/resultsBlock.module.scss
@@ -1,0 +1,61 @@
+@import '../../styles/settings/index';
+
+.title {
+  margin: 0 auto 24px auto;
+  text-align: center;
+  font-size: 24px;
+
+  span {
+    color: $red;
+    font-size: 28px;
+  }
+}
+
+.loader {
+  @include sizes(120px, 120px);
+  margin: 0 auto;
+}
+
+.none {
+  font-size: 32px;
+  line-height: 48px;
+  margin-top: 48px;
+
+  span {
+    color: $red;
+  }
+}
+
+.list {
+  @include grid(5);
+  grid-gap: 16px;
+  margin-bottom: 24px;
+}
+
+@media (max-width: $laptop) {
+  .list {
+    @include grid(4);
+    margin-bottom: 20px;
+  }
+
+  .title {
+    margin-bottom: 20px;
+  }
+}
+
+@media (max-width: $mob) {
+  .list {
+    @include grid(2);
+    margin-bottom: 16px;
+    grid-gap: 12px;
+  }
+
+  .title {
+    font-size: 20px;
+    margin-bottom: 16px;
+
+    span {
+      font-size: 22px;
+    }
+  }
+}
