+++ src/components/ResultsBlock/types.ts
@@ -1,0 +1,8 @@
+import { ICardProps } from '../Card/types';
+
+interface IResultBlockProps {
+  cards: ICardProps[];
+  isLoading: boolean;
+}
+
+export type { IResultBlockProps };
