+++ src/styles/components/_results.scss
@@ -1,36 +1,4 @@
 .results {
-  &__list {
-    @include grid(5);
-    grid-gap: 16px;
-    margin-bottom: 24px;
-  }
-
-  &__title {
-    margin: 0 auto 24px auto;
-    text-align: center;
-    font-size: 24px;
-
-    span {
-      color: $red;
-      font-size: 28px;
-    }
-  }
-
-  &__none {
-    font-size: 32px;
-    line-height: 48px;
-    margin-top: 48px;
-
-    span {
-      color: $red;
-    }
-  }
-
-  &__loader {
-    @include sizes(120px, 120px);
-    margin: 0 auto;
-  }
-
   &__error-btn {
     background-color: $red;
     border-radius: 12px;
@@ -42,38 +10,6 @@
 
     &:hover {
       opacity: 0.8;
-    }
-  }
-}
-
-@media (max-width: $laptop) {
-  .results {
-    &__list {
-      @include grid(4);
-      margin-bottom: 20px;
-    }
-
-    &__title {
-      margin-bottom: 20px;
-    }
-  }
-}
-
-@media (max-width: $mob) {
-  .results {
-    &__list {
-      @include grid(2);
-      margin-bottom: 16px;
-      grid-gap: 12px;
-    }
-
-    &__title {
-      font-size: 20px;
-      margin-bottom: 16px;
-
-      span {
-        font-size: 22px;
-      }
     }
   }
 }
